feat: distribute content to separate modules which are imported to App.tsx  to make it more readable
