--- src/content.txt
@@ -1,38 +1,0 @@
-// nav
-    About me
-
-    Projects
-
-    Contact
-
-// intro page (=about me)
-
-    Hello world
-
-    My name is Jan. I write code, play music and make charts.
-
-    <nav buttons>
-
-    Work experience
-        - Daytrip
-        - IPG
-    
-    Studies
-        - Development economics
-        - Kuwait
-        - Multimedia
-        - International relations
-    
-// Projects
-
-    Spotify dashboard
-        - link to the project 
-        - explain the architecture
-
-// Contact
-
-    simple contact form
-        - contact of person (mandatory)
-        - message
-        - button -> send message to email or telegram
-
